[pt-vulkan][ez] Replace `ArrayRef` with `std::vector<T>&`
Pull Request resolved: https://github.com/pytorch/pytorch/pull/117179

## Context

This change is part of a set of changes that removes all references to the `c10` library in the `api/`, `graph/`, and `impl/` folders of the PyTorch Vulkan codebase. This is to ensure that these components can be built as a standalone library such that they can be used as the foundations of a Android GPU delegate for ExecuTorch.

## Notes for Reviewers

This changeset replaces all instances of `c10::ArrayRef<T>` with `std::vector<T>&` and all instances of`c10::IntArrayRef` with `std::vector<int64_t>&`. There are a lot of changes in this changeset but that is simply due to the large number of callsites. All the changes are straightforward replacements.
ghstack-source-id: 211900526
@exported-using-ghexport

Differential Revision: [D52662235](https://our.internmc.facebook.com/intern/diff/D52662235/)