[pt-vulkan][ez] Replace `c10::overflows` with native equivalent
Pull Request resolved: https://github.com/pytorch/pytorch/pull/117180

## Context

This change is part of a set of changes that removes all references to the `c10` library in the `api/`, `graph/`, and `impl/` folders of the PyTorch Vulkan codebase. This is to ensure that these components can be built as a standalone library such that they can be used as the foundations of a Android GPU delegate for ExecuTorch.


## Notes for Reviewers

This changeset is very straightforward, as it simply copies the required components of `c10::overflows` from [`c10/util/Half.h`](https://github.com/pytorch/pytorch/blob/main/c10/util/Half.h#L477) into `api/Utils.h`.
ghstack-source-id: 211900527
@exported-using-ghexport

Differential Revision: [D52662236](https://our.internmc.facebook.com/intern/diff/D52662236/)