[pt-vulkan] Replace `c10::ScalarType` with native equivalent
Pull Request resolved: https://github.com/pytorch/pytorch/pull/117181

## Context

This change is part of a set of changes that removes all references to the `c10` library in the `api/`, `graph/`, and `impl/` folders of the PyTorch Vulkan codebase. This is to ensure that these components can be built as a standalone library such that they can be used as the foundations of a Android GPU delegate for ExecuTorch.

## Notes for Reviewers

This changeset introduces `api::ScalarType` in `api/Types.h`, which is intended to function the same as `c10::ScalarType`; thus `api/Types.h` is the primary file of interest. The rest of the changes are straightforward replacements of `c10::ScalarType` with `api::ScalarType`.
ghstack-source-id: 211900528
@exported-using-ghexport

Differential Revision: [D52662237](https://our.internmc.facebook.com/intern/diff/D52662237/)