[pt-vulkan][ez] Remove reference to c10::MemoryFormat from `api/` folder
Pull Request resolved: https://github.com/pytorch/pytorch/pull/117183

## Context

This change is part of a set of changes that removes all references to the `c10` library in the `api/`, `graph/`, and `impl/` folders of the PyTorch Vulkan codebase. This is to ensure that these components can be built as a standalone library such that they can be used as the foundations of a Android GPU delegate for ExecuTorch.

## Notes for Reviewers

This changeset removes references to `c10::MemoryFormat` in `api/Tensor.[h,cpp]`; when constructing a `vTensor`, the `api::StorageType` (i.e. whether the tensor will be backed by buffer or texture storage) and `api::GPUMemoryLayout` (i.e. which dimension will be the fastest moving dimension) must be specified directly.
ghstack-source-id: 211900530
@exported-using-ghexport

Differential Revision: [D52662234](https://our.internmc.facebook.com/intern/diff/D52662234/)