Update initializer path for ONNXProgram.save due to onnx.checker limitation
According to https://github.com/onnx/onnx/blob/main/docs/ExternalData.md#large-models-2gb
when initializers are larger than 2GB, `onnx.checker` requires the model
to be in the same directory as the initializer.

Although not strictly necessary for the ONNX export and save to
succeed, it is desirable to have the `onnx.checker` to succeed when
validating the resulting large model.

ghstack-source-id: af1fe0139c72becf1c759d00dbe21041d50097d0
Pull Request resolved: https://github.com/pytorch/pytorch/pull/117294