Add default parameters to rrelu_with_noise()
Summary:
rrelu_with_noise() was listed as having default parameters in the schema but the
actual code definition didn't have them.

The failing example was calling rrelu() which DOES have default parameters and
it passes those defaulted values to C++. Under the covers the C code was calling
the python version of rrelu_with_noise().

Although the C++ code was passing all the values to the python version of
rrelu_with_noise() the pytorch C++ -> Python dispatch code looks at the schema
and strips any parameters which match the schema's listed defaults so if the
schema shows defaults that aren't in the code it will be a problem.

Test Plan:
I added a unit test for this specific case. It would probably be better to write
a more general one to test all the ops - but I haven't learned enough about the
test harness to do that yet.

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: b51f560a66c1ede2d35793284073f00fde19528b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/117141