Ensure that deleter is called even for a no-data tensor.
Summary:

When using a custom deleter InefficientStdFunctionContext was using a
std::unique_ptr<> to store the pointer and call the deleter - but this failed to
call the deleter if the pointer was null. Since we have a separate holder class
anyway take out the std::unique_ptr<> and call the deleter directly.

Fixes #117273

Test Plan:

ghstack-source-id: f0b01281fb37d9af7f5fca0c408d43c49dcd4440
Pull Request resolved: https://github.com/pytorch/pytorch/pull/117418